backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai
langchain-community==0.0.10
chromadb==0.4.22
pypdf==3.17.4
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
